:twisted_rightwards_arrows: Refactor remaining refreshList into getLists
