+++ src/Pages/Common/updateDrugLog.ts
@@ -1,6 +1,6 @@
 import {DrugLogRecord} from "../../types/RecordTypes";
-import RefreshMedicineLog from "../../providers/helpers/RefreshMedicineLog";
 import MedHistoryProvider from "../../providers/MedHistoryProvider";
+import getMedicineLog from "./getMedicineLog";
 
 /**
  * Adds or updates the MedicineHistory table with the given drugLogInfo record
@@ -14,7 +14,7 @@ import MedHistoryProvider from "../../providers/MedHistoryProvider";
 export const updateDrugLog = (medHistoryProvider: typeof MedHistoryProvider, drugLogInfo: DrugLogRecord, residentId: number) => {
     return medHistoryProvider.post(drugLogInfo)
     .then(() => {
-        return RefreshMedicineLog(medHistoryProvider, residentId)
+        return getMedicineLog(medHistoryProvider, residentId)
         .then((drugLogList) => {
             return drugLogList;
         });
